requirements.txt
apache-airflow-providers-apache-spark==5.3.3
pyspark==3.5.7
yfinance==0.2.66